Prevent NPE when sideContainers are not defined in config.xml
When migrating from a previous version of the plugin, or when generating
`config.xml` using a tool such as
[JJB](http://docs.openstack.org/infra/jenkins-job-builder/), not all
configuration elements may be present in the XML. In the case of the
`sideContainers` property, this leads to a `NullPointerException` when
launching a build.

This change forces the deserialized data to be run again through the
`JobBuildsContainersDefinition` constructor, correctly enforcing
the class invariants (such as a non-null `sideContainers` property).

A test was added to expose the unwanted behavior and demonstrate the fix.