Rely on docker exec instead of creating new container for each command to run
As it's still not possible to set the environment with docker exec, we rely
on `trampoline` for setting the working directory and `env` for environment
variables.

This improves pipeline support a lot, as durable task implementation
currently rely on executing a `ps` command every second to check if
the started process is still running.